backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
moviepy 
openai-whisper 
transformers 
sentencepiece 
torch 
ffmpeg-python